+++ dlib/statistics/lda.h
@@ -24,7 +24,7 @@ namespace dlib
             std::map<unsigned long,unsigned long> class_labels;
             for (size_t i = 0; i < row_labels.size(); ++i)
             {
-                const unsigned long next = static_cast<unsigned long>(class_labels.size());
+                const unsigned long next = class_labels.size();
                 if (class_labels.count(row_labels[i]) == 0)
                     class_labels[row_labels[i]] = next;
             }
